fix: support rspack
