+++ packages/core/package.json
@@ -79,7 +79,6 @@
     "@rspack/core": "0.3.6",
     "@rspack/dev-client": "0.3.6",
     "@rspack/plugin-html": "0.3.6",
-    "@rspack/plugin-react-refresh": "0.3.6",
     "@rsbuild/shared": "workspace:*",
     "@svgr/webpack": "8.0.1",
     "commander": "^10.0.1",
