+++ packages/plugin-react/package.json
@@ -24,6 +24,7 @@
   ],
   "license": "MIT",
   "dependencies": {
+    "@rspack/plugin-react-refresh": "0.3.6",
     "@babel/preset-react": "^7.22.15",
     "@modern-js/utils": "^2.37.1",
     "@rsbuild/shared": "workspace:*",
