+++ packages/plugin-react/tests/rspack.test.ts
@@ -1,0 +1,52 @@
+import { expect, describe, it } from 'vitest';
+import { pluginReact } from '../src';
+import { createStubRsbuild } from '@rsbuild/test-helper';
+
+describe('plugins/react', () => {
+  it('should work with swc-loader', async () => {
+    const rsbuild = await createStubRsbuild({
+      rsbuildConfig: {},
+    });
+
+    rsbuild.addPlugins([pluginReact()]);
+    const config = await rsbuild.unwrapConfig();
+
+    expect(config).toMatchSnapshot();
+  });
+
+  it('should not apply react refresh when dev.hmr is false', async () => {
+    const rsbuild = await createStubRsbuild({
+      rsbuildConfig: {
+        dev: {
+          hmr: false,
+        },
+      },
+    });
+
+    rsbuild.addPlugins([pluginReact()]);
+
+    expect(await rsbuild.matchBundlerPlugin('ReactRefreshPlugin')).toBeFalsy();
+  });
+
+  it('should not apply react refresh when target is node', async () => {
+    const rsbuild = await createStubRsbuild({
+      target: 'node',
+      rsbuildConfig: {},
+    });
+
+    rsbuild.addPlugins([pluginReact()]);
+
+    expect(await rsbuild.matchBundlerPlugin('ReactRefreshPlugin')).toBeFalsy();
+  });
+
+  it('should not apply react refresh when target is web-worker', async () => {
+    const rsbuild = await createStubRsbuild({
+      target: 'web-worker',
+      rsbuildConfig: {},
+    });
+
+    rsbuild.addPlugins([pluginReact()]);
+
+    expect(await rsbuild.matchBundlerPlugin('ReactRefreshPlugin')).toBeFalsy();
+  });
+});
